requirements.txt
pyyaml==5.3.1
awsebcli==3.20.3
backports.entry-points-selectable==1.1.1
backports.shutil-get-terminal-size==1.0.0
blinker==1.5
botocore==1.23.54
cement==2.8.2
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.1
colorama==0.4.3
comtypes==1.1.14
configparser==5.0.2
cryptography==38.0.1
Deprecated==1.2.13
distlib==0.3.6
dnspython==2.2.1
enum34==1.1.10
filelock==3.8.0
flake8==3.9.2
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
frozenlist==1.3.3
future==0.16.0
gcloud==0.18.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.2
importlib-metadata==4.12.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jwcrypto==1.4.2
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==6.0.4
mysql-connector==2.2.9
mysql-connector-python==8.0.31
nose==1.3.7
oauth2client==4.1.3
openai==0.27.0
pymongo==4.2.0
pyTelegramBotAPI==4.9.0
python-dateutil==2.8.1
python-dotenv==0.21.0
virtualenv==20.8.1
virtualenv-clone==0.5.7
Werkzeug==2.0.1
wrapt==1.14.1
yarl==1.8.2
zipp==3.8.1
telethon==1.26.1
pyopenssl==23.0.0
twikit
replicate
python-dotenv
mistralai
tweepy